Add interim-update script to pppd_compat
When the delegated IPv6 prefix is not set by radius beforehand, the ip-up script will not be able to handle the prefix, as the DHCPv6 stuff is happening way after the ip-up script.
To allow some sort of script to handle the dynamic delegation of a prefix, add an interim-update script.
An example scenario would be the addition of some firewall rules for the prefix.

This script is called exactly the same as the ip-up script.