ipoe: add an option to not send option 82 to clients
DHCP option 82 fields carry the DHCP relay agent information to the DHCP
servers. Option 82 fields do not normally need to be disclosed to the
DHCP clients, unless the client itself is a relay.

Add an option to not send DHCP option 82 fields to clients. By default,
the fields are sent to not change the current behavior.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>