ipoe: retry connecting to the DHCP server
At startup, accel-pppd may be unable to create the UDP socket to the
DHCP server because the route to the destination does not yet exist.
It is common issue when the routes are fetched from a dynamic routing
protocol.

> Jun 20 09:56:16 RTR-RELAY accel-pppd[618]: dhcpv4: relay: 10.90.81.68: connect: No route to host

As a result, accel-pppd is unable to relay the DHCP packets from the
hosts.

> Jun 20 10:11:58 RTR-RELAY accel-pppd[614]: vxl3.333:: recv [DHCPv4 Discover xid=55d9df95 chaddr=0c:cc:2a:94:00:00 <Message-Type Discover> <Request-IP 10.101.0.32> <Host-Name h6> <Request-List Subnet,Broadcast,Time-Offset,Router,Domain-Name,DNS,119,Host-Name,44,47,MTU,Classless-Route,NTP> <Client-ID ff2a940000000100012c1d92cf0ccc2a940000>]
> Jun 20 10:11:58 RTR-RELAY accel-pppd[614]: ipoe0:: create interface ipoe0 parent vxl3.333
> Jun 20 10:11:58 RTR-RELAY accel-pppd[614]: ipoe0:: (null): authentication succeeded
> Jun 20 10:11:58 RTR-RELAY accel-pppd[614]: ipoe0:: no free IPv4 address
< Jun 20 10:11:58 RTR-RELAY accel-pppd[614]: ipoe0:: terminate
> Jun 20 10:11:58 RTR-RELAY accel-pppd[614]: ipoe0:: ipoe: session finished

Retry to connect the socket if it was not set after receiving a DHCP
packet from a client.

Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>