ipoe: fix session termination with unresponsive DHCP server
If the DHCP server stops replying to the relay, the session never
terminates unless the client stops sending DHCP REQUEST. The lease
timeout is supposed to terminate the session if the lease is not renewed
after the lease time that is specified by the DHCP server or the
configured max-lease-time. However, the keepalive timer is reset each
time a DHCP REQUEST is received from the client.

Reset the timer each time a DHCP ACK from the server is relayed to the
client instead of doing it after receiving a DHCP REQUEST.

Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>