ipoe: set siaddr to giaddr at last resort
siaddr is the "IP address of next server to use in bootstrap returned in
DHCPOFFER, DHCPACK by server".

siaddr is set at last resort to serv->dhcpv4_relay->giaddr that is the
configured giaddr or else the source address of the serv->dhcpv4_relay
socket. However serv->dhcpv4_relay will disappear in favor of a list of
dhcpv4_relay.

Set siaddr to the first found giaddr value in the list.

Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>