ipoe: relay requests to the relevant servers only
Accel-ppp now forwards all the client requests to all the configured
DHCP servers. However, only the server that provides the DHCP parameters
is interested in receiving all the requests.

At the lease initialization, the DHCP client sends a DHCPDISCOVER that
MUST be relayed to all the configured servers. Multiple DHCPOFFER are
received in return. The accel-ppp relay agent only considers the first
one and relays it to client. The client sends a first DHCPREQUEST with
no filled client IP address (ciaddr) but with a "Requested IP address"
(DHCP option 50). This DHCPREQUEST MUST be relayed to all the servers
with the "server ID" (option 54) of the first DHCPOFFER. The matching
server knows its offer is accepted and the others implicitly knows their
offer is declined. The next DHCP requests from the client are only
relevant for the matching.

After the lease initialization, relay requests to the server that
provides the first DHCPOFFER only.

Link: https://www.rfc-editor.org/rfc/rfc2131#page-16
Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>