ipoe: relay client requests to servers that provided yiaddr
Several DHCP servers may provide a similar DHCPOFFER. accel-ppp
only considers the first received offer. The initial DHCPREQUEST is
relayed with the "server ID" (option 54) of the first DHCPOFFER
to all the servers. Theoretically, only the matching server should reply
with a DHCPACK. However, some implementation like ISC DHCP server
ignores the "server ID" (option 54) value and replies with a DHCPACK
if the requested options match the DHCPOFFER it sent.

Allow relaying the requests to all DHCP that provides the session
yiaddr.

Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>