ipoe: fix echoing back dhcp option 82
RFC3046 says:

> The Relay Agent Information option [82] echoed by a server MUST be removed
> by either the relay agent or the trusted downstream network element
> which added it when forwarding a server-to-client response back to
> the client.

985cb9f994 ("ipoe: dhcpv4: echo back opt82 if sent by client/unknown
relay per rfc3046") was supposed to implement this requirement. However,
it does not work because dhcpv4_relay_send() modifies the original
incoming request before relaying it.

Fix the requirement conformity by working on a copy of the original
request in dhcpv4_relay_send().

Tested with the dhtest tool with and without an option 82:

> ./dhtest -i ens3
> ./dhtest -c 82,hex,010876786c312e31313102104147454e542d52454d4f54452d49443105040a0f4002 -i ens3

Fixes: 985cb9f994 ("ipoe: dhcpv4: echo back opt82 if sent by client/unknown relay per rfc3046")
Link: https://www.rfc-editor.org/rfc/rfc3046.html#section-2.1
Link: https://github.com/saravana815/dhtest
Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>