ipoe: add a trusted-circuit option
RFC3046 distinguishes "trusted" and "untrusted" circuits from which a
DHCP request can come:

> The mechanisms for distinguishing between "trusted" and "untrusted"
> circuits are specific to the type of circuit termination equipment,
> and may involve local administration.  For example, a Cable Modem
> Termination System may consider upstream packets from most cable
> modems as "untrusted", but an ATM switch terminating VCs switched
> through a DSLAM may consider such VCs as "trusted" and accept a relay
> agent option added by the DSLAM.

Add a global and per interface trusted-circuit configuration option to
define whether the agent information, such as the giaddr and the option
82, in incoming DHCP requests should be trusted.

Security will be enforced on next commits on untrusted interfaces.
Default the value to 1 (trust) to keep the current behavior.

Link: https://www.rfc-editor.org/rfc/rfc3046.html#section-2.1
Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>