feat(build): Add MUSL detection and conditional linking
This commit introduces the ability to detect if the project is being
built with the MUSL C library.

A new variable `MUSL` is set to `ON` if MUSL is detected, and `OFF`
otherwise. This is achieved by checking the output of `ldd --version`.

The `accel-pppd/ctrl/pppoe/CMakeLists.txt` file is updated to
conditionally link the `connlimit` library only when building with MUSL.

Signed-off-by: Denys Fedoryshchenko <denys.f@collabora.com>